Full rewrite. Previous `out()` design didn't account for build-time plugins
I missed the fine print in the Rollup documentation and didn't understand the
pipeline; @rollup/plugin-babel and @rollup/plugin-replace (before I rewrote it)
cannot be output-time plugins. However, out() generated format-specific configs
under _one_ bundle which made it difficult to then go back in and split ESM vs
UMD+IIFE bundles separately. Not only in terms of code complexity but also the
readability and understanding: what would it depend on? would it be hardcoded?

This clears all of that up. Also the entire document reads better top to bottom
and makes use of function hoisting to keep it clean.

I replaced the replace plugin with something more powerful to allow for more
complicated optimizations (such as undoing transpilation in specific parts of a
minified bundle where the exact string to replace can't be known ahead of time)